Struct mdrc_pacbot_util::game_state::PacmanState

pub struct PacmanState {

    mode: GhostMode,
    old_mode: GhostMode,
    just_swapped_state: bool,
    state_counter: u32,
    start_counter: u32,
    pub paused: bool,
    pub score: usize,
    frightened_counter: u8,
    pub frightened_multiplier: u8,
    pub lives: u8,
    pub elapsed_time: u32,
    pub pacman: Agent,
    pub ghosts: Vec<Ghost>,
    pub pellets: Vec<bool>,
    pub power_pellets: Vec<Point2<u8>>,

}

Information that changes during a game of Pacman

Note: frightened_counter is not present because its only effect is Pacman’s speed after collecting a power pellet

Fields
mode: GhostMode

Current ghost behavior - applies to all ghosts

When paused, Pacman should not move

old_mode: GhostMode

Mode before the super pellet

just_swapped_state: bool

Whether we entered frightened mode or swapped states the previous tick

state_counter: u32

Determines when the pre-programmed state swaps happen

start_counter: u32

Determines how ghosts follow their starting paths

paused: bool

Whether the game is paused

score: usize

Player’s current game score

frightened_counter: u8

Global time remaining for ghosts to be frightened

frightened_multiplier: u8

Bonus for capturing multiple ghosts in a power pellet

lives: u8

Lives remaining - starts at 3; at 0, the game is over

elapsed_time: u32

Number of frames that have passed since the start of the game

pacman: Agent

Pacman’s location and direction

ghosts: Vec<Ghost>

Ghosts

pellets: Vec<bool>

Pellets remaining

power_pellets: Vec<Point2<u8>>

Super pellets remaining

Implementations

impl PacmanState

pub fn new(agent_setup: &PacmanAgentSetup) -> Self

Create a new PacmanState from a PacmanAgentSetup

pub fn reset(&mut self, agent_setup: &PacmanAgentSetup, use_physics: bool)

Reset the game state to the initial state using the same or different PacmanAgentSetup

pub fn respawn_agents( &mut self, agent_setup: &PacmanAgentSetup, use_physics: bool )

Respawn the ghosts and Pacman, for when Pacman dies

pub fn step( &mut self, agent_setup: &PacmanAgentSetup, rng: &mut ThreadRng, use_physics: bool )

Move forward one frame, using the current Pacman location

pub fn update_pacman(&mut self, p: Point2<u8>, d: Direction)

Update Pacman’s location and direction

pub fn pause(&mut self)

Pause the game

pub fn resume(&mut self)

Resume the game

fn is_game_over(&self) -> bool

Test if the game is over (if all pellets are eaten)

fn should_die(&self, use_physics: bool) -> bool

Should Pacman die this step?

fn die(&mut self, agent_setup: &PacmanAgentSetup, use_physics: bool)

Pacman dies

fn check_if_ghost_eaten(&mut self, agent_setup: &PacmanAgentSetup)

fn update_ghosts(&mut self, agent_setup: &PacmanAgentSetup, rng: &mut ThreadRng)

fn update_score(&mut self, grid: &ComputedGrid)