Struct mdrc_pacbot_util::gui::physics::PhysicsRenderInfo

pub struct PhysicsRenderInfo {
    pub sleep: bool,
    pub pacbot_pos: Isometry2<f32>,
    pub pacbot_pos_guess: Isometry2<f32>,
    pub primary_robot_rays: Vec<(Point2<f32>, Point2<f32>)>,
    pub pf_count: usize,
    pub pf_points: Vec<Isometry2<f32>>,
}

Stores state needed to render physics information.

Fields
sleep: bool

If true, the physics thread should not advance physics

pacbot_pos: Isometry2<f32>

The current position of the robot.

pacbot_pos_guess: Isometry2<f32>

The particle filter’s current best guess

primary_robot_rays: Vec<(Point2<f32>, Point2<f32>)>

An array of start and end points.

pf_count: usize

The number of best particle filter points to save

pf_points: Vec<Isometry2<f32>>

The best pf_count particle filter points