Struct mdrc_pacbot_util::physics::particle_filter::ParticleFilter

pub struct ParticleFilter {
    robot: Robot,
    grid: ComputedGrid,
    raycast_grid: RaycastGrid,
    points: Vec<FilterPoint>,
    best_guess: FilterPoint,
    options: ParticleFilterOptions,
}

Tracks the robot’s position over time

Fields
robot: Robot

Robot specifications

grid: ComputedGrid

The grid used to find empty spaces; to change this, create a new particle filter

raycast_grid: RaycastGrid

The data structure for performing raycasts on the physical grid.

points: Vec<FilterPoint>

Guesses for the current location, ordered by measured accuracy

best_guess: FilterPoint

The current best guess

options: ParticleFilterOptions

Values that can be tweaked to improve the performance of the particle filter

Implementations

impl ParticleFilter

pub fn new( grid: ComputedGrid, robot: Robot, start: Isometry2<f32>, options: ParticleFilterOptions ) -> Self

Create a ParticleFilter

Start determines the location around which the filter will generate initial particles

pub fn set_options(&mut self, particle_filter_options: ParticleFilterOptions)

pub fn set_robot(&mut self, robot: Robot)

pub fn raycast_grid(&self) -> &RaycastGrid

fn random_point_uniform(&self) -> Isometry2<f32>

Generate a completely random walkable point

fn random_point_at(&self, node: IntLocation, rng: ThreadRng) -> Isometry2<f32>

Generate a random valid point around a certain walkable square

pub fn update( &mut self, velocity: Isometry2<f32>, dt: f32, stopwatch: &mut Stopwatch, sensors: &PacbotSensors, settings: &UserSettings )

Update the particle filter, using the same rigid body set as the start

fn distance_sensor_diff( &self, robot: &Robot, point: Isometry2<f32>, actual_values: &Vec<Option<f32>> ) -> f32

Given a location guess, measure the absolute difference against the real values

fn distance_sensor_ray( &self, point: Isometry2<f32>, sensor: DistanceSensor ) -> f32

Given a location guess, measure one sensor

pub fn points(&self, count: usize) -> Vec<FilterPoint>

Get the best ‘count’ particle filter points

pub fn best_guess(&self) -> FilterPoint

Get the best guess