Struct mdrc_pacbot_util::physics::PacbotSimulation

pub struct PacbotSimulation {

    integration_parameters: IntegrationParameters,
    physics_pipeline: PhysicsPipeline,
    island_manager: IslandManager,
    broad_phase: BroadPhase,
    narrow_phase: NarrowPhase,
    impulse_joint_set: ImpulseJointSet,
    multibody_joint_set: MultibodyJointSet,
    ccd_solver: CCDSolver,
    rigid_body_set: RigidBodySet,
    collider_set: ColliderSet,
    query_pipeline: QueryPipeline,
    query_pipeline_updated: bool,
    robot_specifications: Robot,
    primary_robot: ColliderHandle,
    robot_target_velocity: (Vector2<f32>, f32),
    particle_filter: ParticleFilter,

}

Handles all physics related operations

Fields
integration_parameters: IntegrationParametersphysics_pipeline: PhysicsPipelineisland_manager: IslandManagerbroad_phase: BroadPhasenarrow_phase: NarrowPhaseimpulse_joint_set: ImpulseJointSetmultibody_joint_set: MultibodyJointSetccd_solver: CCDSolverrigid_body_set: RigidBodySetcollider_set: ColliderSetquery_pipeline: QueryPipelinequery_pipeline_updated: boolrobot_specifications: Robotprimary_robot: ColliderHandlerobot_target_velocity: (Vector2<f32>, f32)particle_filter: ParticleFilter
Implementations

impl PacbotSimulation

pub fn pf_update( &mut self, position: Point2<u8>, pf_stopwatch: &Arc<Mutex<Stopwatch>> )

Update the particle filter

impl PacbotSimulation

pub fn new( grid: ComputedGrid, robot: Robot, robot_position: Isometry2<f32>, distance_sensors: Arc<Mutex<Vec<Option<f32>>>> ) -> Self

Create a new simulation on a ComputedGrid with a starting Robot and position

Examples

use std::sync::{Arc, Mutex};
use rand::rngs::ThreadRng;
use rapier2d::na::{Isometry2, Vector2};
use mdrc_pacbot_util::grid::ComputedGrid;
use mdrc_pacbot_util::robot::Robot;
use mdrc_pacbot_util::physics::PacbotSimulation;
use mdrc_pacbot_util::standard_grids::StandardGrid;

let grid = StandardGrid::Pacman.compute_grid();
let robot = Robot::default();
let distance_sensors = Arc::new(Mutex::new(vec![Some(0.0); Robot::default().distance_sensors.len()]));
let starting_position = Isometry2::new(Vector2::new(14.0, 7.0), 0.0);
let mut simulation = PacbotSimulation::new(grid, robot, starting_position, distance_sensors);

pub fn step(&mut self)

Update the physics simulation

Examples

use mdrc_pacbot_util::physics::PacbotSimulation;
let mut simulation = PacbotSimulation::default();

// in an infinite loop
simulation.step();

fn step_target_velocity(&mut self)

Apply an impulse to the primary robot based on robot_target_velocity

pub fn get_collider_position( &self, handle: ColliderHandle ) -> Option<&Isometry2<f32>>

Get the Isometry for a given [ColliderHandle]

May return None under any of the following conditions:

there is no matching collider

the collider in question has no associated rigid body

the rigid body associated with the collider is invalid

Examples

use rapier2d::geometry::ColliderHandle;
use rapier2d::na::{Isometry2, Point2, Vector2};
use rapier2d::prelude::Rotation;
use mdrc_pacbot_util::physics::PacbotSimulation;
let mut simulation = PacbotSimulation::default();

// in an infinite loop
let collider: ColliderHandle = simulation.get_primary_robot_collider();
let isometry: &Isometry2<f32> = simulation.get_collider_position(collider).unwrap();
let position: Point2<f32> = isometry.translation.transform_point(&Point2::new(0.0, 0.0));
let rotation: Rotation<f32> = isometry.rotation;

pub fn cast_ray(&mut self, ray: Ray, max_toi: Real) -> Point<Real>

Cast a ray in the simulation

Rays will pass through robots and only hit walls. It is recommended to normalize the Ray’s direction so that max_toi represents a maximum distance. If the ray does not strike a wall within max_toi, the point at max_toi along the ray will be returned.

Examples

use rapier2d::na::{Point2, Vector2};
use rapier2d::prelude::{Ray, Rotation};
use mdrc_pacbot_util::physics::PacbotSimulation;
let mut simulation = PacbotSimulation::default();

let pacbot_position = simulation.get_primary_robot_position();
let positive_y = Vector2::new(0.0, 1.0);
let ray = Ray::new(pacbot_position.translation.transform_point(&Point2::new(0.0, 0.0)), positive_y);

assert_eq!(simulation.cast_ray(ray, 5.0), Point2::new(14.0, 8.0));
assert_eq!(simulation.cast_ray(ray, 0.5), Point2::new(14.0, 7.5));

pub fn get_primary_robot_collider(&self) -> ColliderHandle

Get the collider for the primary robot

Examples

use rapier2d::prelude::ColliderHandle;
use mdrc_pacbot_util::physics::PacbotSimulation;
let mut simulation = PacbotSimulation::default();

let primary_collider: ColliderHandle = simulation.get_primary_robot_collider();

pub fn get_primary_robot_position(&self) -> &Isometry2<f32>

Get the current position of the primary robot

Examples

use rapier2d::math::Rotation;
use rapier2d::na::Point2;
use mdrc_pacbot_util::physics::PacbotSimulation;
let mut simulation = PacbotSimulation::default();

let isometry = simulation.get_primary_robot_position();
let position: Point2<f32> = isometry.translation.transform_point(&Point2::new(0.0, 0.0));
let rotation: Rotation<f32> = isometry.rotation;

pub fn set_target_robot_velocity(&mut self, v: (Vector2<f32>, f32))

Set the target velocity (translational and rotational) for the primary robot

Examples

use rapier2d::na::Vector2;
use mdrc_pacbot_util::physics::PacbotSimulation;
let mut simulation = PacbotSimulation::default();

let w_key_pressed = true;
if w_key_pressed {
    simulation.set_target_robot_velocity((Vector2::new(0.0, 1.0), 0.0));
}

pub fn get_primary_robot_rays(&mut self) -> Vec<(Point<Real>, Point<Real>)>

Get the rays coming out of the primary robot, representing the theoretical readings from its distance sensors.

Examples

use mdrc_pacbot_util::robot::Robot;
use mdrc_pacbot_util::physics::PacbotSimulation;
let mut simulation = PacbotSimulation::default();

let robot = Robot::default();
let rays = simulation.get_primary_robot_rays();

for i in 0..robot.distance_sensors.len() {
    let sensor = robot.distance_sensors[i];

    let sensor_position = rays[i].0;
    let hit_point = rays[i].1;

    let difference = hit_point - sensor_position;

    assert_eq!((sensor.relative_direction.cos() >= 0.1), (difference.x >= 0.1));
    assert_eq!((sensor.relative_direction.sin() >= 0.1), (difference.y >= 0.1));
}

pub fn pf_best_guess(&self) -> Isometry2<f32>

Get the particle filter’s best guess position

pub fn pf_points(&self, count: usize) -> Vec<Isometry2<f32>>

Get the best ‘count’ particle filter points