Struct mdrc_pacbot_util::UserSettings

pub struct UserSettings {

    pub mode: AppMode,
    pub enable_ai: bool,
    pub pico_address: Option<String>,
    pub go_server_address: Option<String>,
    pub robot: Robot,
    pub sensors_from_robot: bool,
    pub kidnap_position: Option<IntLocation>,
    pub replay_save_location: bool,
    pub replay_save_sensors: bool,
    pub replay_save_targets: bool,
    pub enable_pf: bool,
    pub pf_total_points: usize,
    pub pf_gui_points: usize,
    pub pf_error_threshold: f32,
    pub pf_chance_near_other: f32,
    pub pf_avg_kidnaps_per_sec: f32,
    pub pf_cv_error_std: f32,
    pub pf_translation_limit: f32,
    pub pf_rotation_limit: f32,
    pub pf_simulated_translation_noise: f32,
    pub pf_simulated_rotation_noise: f32,
    pub pf_generic_noise: f32,

}

Options that the user can set via the GUI, shared between most processes

Fields
mode: AppMode

Whether the app is recording (normal) or playback for a replay

enable_ai: bool

Whether AI actions should be calculatede

pico_address: Option<String>

Optional IP for the pico

go_server_address: Option<String>

Optional IP for the game server

robot: Robot

Physical characteristics of the robot

sensors_from_robot: bool

Whether sensor values should come from the robot, versus rapier

kidnap_position: Option<IntLocation>

When the user clicks on a location where the simulated robot should be teleported

replay_save_location: bool

Whether physical location should be saved in the replay

replay_save_sensors: bool

Whether pacbot sensors should be saved in the replay

replay_save_targets: bool

Whether target paths and velocities should be saved in the replay

enable_pf: bool

Whether particle filter is calculated

pf_total_points: usize

The number of guesses tracked by ParticleFilter

pf_gui_points: usize

The number of points displayed on the gui

pf_error_threshold: f32

All points with a larger error are removed

pf_chance_near_other: f32

Chance 0.0-1.0 that a new point will spawn near an existing one instead of randomly

pf_avg_kidnaps_per_sec: f32

The average number of times the robot is kidnapped per second, in our theoretical motion model. This determines the probability that a particle will be teleported to a random position.

pf_cv_error_std: f32

The standard deviation of the CV position error, in our theoretical sensor model.

pf_translation_limit: f32

When generating a point based on an existing point, how far can it be moved in x and y?

pf_rotation_limit: f32

When generating a point based on an existing point, how far can it be moved in rotation?

pf_simulated_translation_noise: f32

When moving particles by Rapier-reported distance, add noise proportional to translation

pf_simulated_rotation_noise: f32

When moving particles by Rapier-reported distance, add noise proportional to rotation

pf_generic_noise: f32

When moving particles by Rapier-reported distance, add noise